requirements.txt
discord.py
python-dotenv
pillow
pydub
aiohttp
pytz
numpy